requirements.txt
openai>=1.3.9
crewai>=0.26.9
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
python-dotenv>=1.0.1
pillow>=10.3.0
scikit-learn>=1.4.2
torch>=2.3.0
torchvision>=0.18.0
transformers==4.39.3